gitupload support for directory
